Moves compliance status to utils
